# Conflicts:
#	package.json
#	public_html/admin/catalog.app/csv.inc.php
#	public_html/admin/customers.app/csv.inc.php
#	public_html/admin/orders.widget/orders.inc.php
#	public_html/admin/pages.app/csv.inc.php
#	public_html/admin/translations.app/csv.inc.php
#	public_html/backend/apps/catalog/catalog.inc.php
#	public_html/backend/apps/catalog/edit_product.inc.php
#	public_html/backend/pages/index.inc.php
#	public_html/frontend/boxes/box_cart.inc.php
#	public_html/frontend/pages/printable_packing_slip.inc.php
#	public_html/frontend/pages/product.inc.php
#	public_html/frontend/templates/default/views/box_checkout_payment.inc.php
#	public_html/frontend/templates/default/views/box_checkout_shipping.inc.php
#	public_html/includes/app_header.inc.php
#	public_html/includes/autoloader.inc.php
#	public_html/includes/entities/ent_image.inc.php
#	public_html/includes/entities/ent_product.inc.php
#	public_html/includes/functions/func_form.inc.php
#	public_html/includes/functions/func_general.inc.php
#	public_html/includes/functions/func_reference.inc.php
#	public_html/includes/modules/order_total/ot_payment_fee.inc.php
#	public_html/includes/nodes/nod_route.inc.php
#	public_html/includes/references/ref_product.inc.php
#	public_html/index.php
#	public_html/install/install.php
#	public_html/install/upgrade.php
#	public_html/pages/ajax/checkout_customer.inc.php
#	public_html/pages/ajax/checkout_payment.inc.php
#	public_html/pages/ajax/checkout_shipping.inc.php
#	public_html/pages/manufacturer.inc.php

+++ public_html/includes/nodes/nod_volume.inc.php
@@ -103,7 +103,7 @@
 
     public static function convert($value, $from, $to) {
 
-      if ($value == 0) return 0;
+      if ((float)$value == 0) return 0;
 
       if ($from == $to) return (float)$value;
 
@@ -117,7 +117,7 @@
         return;
       }
 
-      if (self::$classes[$from]['value'] == 0 || self::$classes[$to]['value'] == 0) return;
+      if ((float)self::$classes[$from]['value'] == 0 || (float)self::$classes[$to]['value'] == 0) return;
 
       return $value * (self::$classes[$to]['value'] / self::$classes[$from]['value']);
     }
